Teraz Już wszystko powinno działać
